Uploaded code & .csv
YES, the .csv has patient IDs from the secured database, but it's not nearly raw: literally just <patientID1, patientID2, JaccardSimilarity> So I figure it's okay if I post it here to save you guys the trouble of running it.  

PS: I tested it by comparing patient 5 w/ patient 9 and it checks out.